+++ CHANGELOG.md
@@ -174,6 +174,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * `setTracking` method to the `Order` entity
+* `setReturnTracking` method to the `Order` entity
 * `setMeta` method allowing dynamic set of metadata attributes for the `Order` entity
 * Method for generating an image URL from a query (`_queryToImageUrl`)